HZN-1182: Make elastic-credentials.xml more generic
