--- features/flows/elastic/src/main/java/org/opennms/netmgt/flows/elastic/credentials/ElasticFlowRepositoryCredentialsProvider.java
@@ -1,60 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2017-2017 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2017 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.netmgt.flows.elastic.credentials;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.xml.bind.JAXB;
-
-import org.opennms.plugins.elasticsearch.rest.credentials.CredentialsProvider;
-import org.opennms.plugins.elasticsearch.rest.credentials.CredentialsScope;
-
-public class ElasticFlowRepositoryCredentialsProvider implements CredentialsProvider {
-
-    private final Path configPath;
-
-    public ElasticFlowRepositoryCredentialsProvider(String configName) {
-        Path configPath = Paths.get(System.getProperty("opennms.home"), "etc", configName);
-        this.configPath = configPath;
-    }
-
-    @Override
-    public List<CredentialsScope> getCredentials() {
-        if (Files.exists(configPath)) {
-            final ElasticFlowRepositoryCredentialsWrapper credentialsWrapper = JAXB.unmarshal(configPath.toFile(),
-                    ElasticFlowRepositoryCredentialsWrapper.class);
-            return credentialsWrapper.getCredentialsScopes();
-        }
-        return new ArrayList<>();
-    }
-}
