+++ opennms-base-assembly/src/main/filtered/etc/elastic-credentials.xml
